+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Sy 2018 etirinotecan pegol ([doi:10.1007/s00280-018-3562-3](https://doi.org/10.1007/s00280-018-3562-3)) -- adults with advanced solid tumors.
+
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.
 
 - Add Nguyen 2026 dupilumab ([doi:10.1002/cpt.70233](https://doi.org/10.1002/cpt.70233)) -- healthy adults and adults, adolescents and children aged 1 year and older with eosinophilic esophagitis.